Fixed event clashes, refactored structure.
